requirements.txt
rich==12.4.4
# numpy==1.19.2
opencv_python==4.5.4.58
scipy==1.5.4
tqdm==4.62.3
PyYAML==6.0
# dgl==0.8.0post2
# open3d==0.15.2
scikit-image==0.19.2
scikit-learn==1.0.2
munch==2.5.0
# torch==1.10.1
ninja==1.10.2.3
matplotlib==3.5.1
tifffile==2021.11.2
imgaug==0.4.0
kornia==0.6.5
torchmetrics==0.9.1
faiss-gpu==1.7.2
timm==0.6.11
efficientnet_pytorch==0.7.1
FrEIA==0.2
ignite==0.4.10
einops==0.4.1
torchprofile==0.0.4
pytorch-msssim==0.2.1